Updating poetry install
The poetry install step in github actions starts installing python modules but falls over on github actions with this error "/home/runner/work/py_in_rpkg/py_in_rpkg/py_in_rpkg does not contain any element
Error: Process completed with exit code 1.".  It appears it's trying to install the project itself and can't find it as a module.

Two options to resolve can be found here -> https://stackoverflow.com/questions/75397736/poetry-install-on-an-existing-project-error-does-not-contain-any-element.

I'm trying the second option of adding "--no--root" to tell it not to install the root project. This avoids making changes to the python project toml document incase it affects development env things.